refactor: [MON-302] 스토리북 폴더 구조 변경 &  패키지 관련 내용 변경됨
