--- src/stories/components/domain/CategoryMenu.stories.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { CategoryMenu } from '@components';
-
-export default {
-  title: 'Component/domain/CategoryMenu',
-  component: CategoryMenu,
-};
-
-export const Default = () => <CategoryMenu />;
